add graphql codegen
