+++ utils/helpers.js
@@ -5,3 +5,11 @@ export const cls = input =>
         .filter(cond => typeof cond === "string")
         .join(" ")
         .trim();
+
+export const truncateDescriptions = (description, limit = 100) => {
+    if (description.length <= limit) {
+        return description;
+    } else {
+        return description.substring(0, limit) + "...";
+    }
+};
